@@ -14,9 +14,9 @@
     "graphql": "^0.5.0",
     "graphql-relay": "^0.4.3",
     "isomorphic-relay": "file:../..",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
-    "react-relay": "0.9.3"
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
+    "react-relay": "^1.7.0"
   },
   "devDependencies": {
     "babel-cli": "^6.3.17",
