@@ -33,8 +33,8 @@
     "babel-runtime": "^6.6.1"
   },
   "peerDependencies": {
-    "react": "^0.14.8 || ^15.0.1",
-    "react-relay": "0.9.3 - 0.10.0"
+    "react": "^16.3.0",
+    "react-relay": "^1.7.0"
   },
   "devDependencies": {
     "babel-cli": "^6.7.5",
@@ -44,7 +44,7 @@
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
     "babel-relay-plugin": "0.9.3",
-    "react": "^15.0.1",
-    "react-relay": "0.9.3"
+    "react": "^16.3.0",
+    "react-relay": "^1.7.0"
   }
 }
